+++ lib/flipper/adapters/active_record.rb
@@ -62,18 +62,20 @@ module Flipper
 
       # Public: The set of known features.
       def features
-        @feature_class.all.map(&:key).to_set
+        with_connection(@feature_class) { @feature_class.all.map(&:key).to_set }
       end
 
       # Public: Adds a feature to the set of known features.
       def add(feature)
-        # race condition, but add is only used by enable/disable which happen
-        # super rarely, so it shouldn't matter in practice
-        @feature_class.transaction do
-          unless @feature_class.where(key: feature.key).first
-            begin
-              @feature_class.create! { |f| f.key = feature.key }
-            rescue ::ActiveRecord::RecordNotUnique
+        with_connection(@feature_class) do
+          # race condition, but add is only used by enable/disable which happen
+          # super rarely, so it shouldn't matter in practice
+          @feature_class.transaction do
+            unless @feature_class.where(key: feature.key).first
+              begin
+                @feature_class.create! { |f| f.key = feature.key }
+              rescue ::ActiveRecord::RecordNotUnique
+              end
             end
           end
         end
@@ -83,16 +85,18 @@ module Flipper
 
       # Public: Removes a feature from the set of known features.
       def remove(feature)
-        @feature_class.transaction do
-          @feature_class.where(key: feature.key).destroy_all
-          clear(feature)
+        with_connection(@feature_class) do
+          @feature_class.transaction do
+            @feature_class.where(key: feature.key).destroy_all
+            clear(feature)
+          end
         end
         true
       end
 
       # Public: Clears the gate values for a feature.
       def clear(feature)
-        @gate_class.where(feature_key: feature.key).destroy_all
+        with_connection(@gate_class) { @gate_class.where(feature_key: feature.key).destroy_all }
         true
       end
 
@@ -100,29 +104,52 @@ module Flipper
       #
       # Returns a Hash of Flipper::Gate#key => value.
       def get(feature)
-        db_gates = @gate_class.where(feature_key: feature.key)
-        result_for_feature(feature, db_gates)
+        gates = with_connection(@gate_class) { @gate_class.where(feature_key: feature.key).pluck(:key, :value) }
+        result_for_gates(feature, gates)
       end
 
       def get_multi(features)
-        db_gates = @gate_class.where(feature_key: features.map(&:key))
-        grouped_db_gates = db_gates.group_by(&:feature_key)
-        result = {}
-        features.each do |feature|
-          result[feature.key] = result_for_feature(feature, grouped_db_gates[feature.key])
+        with_connection(@gate_class) do
+          gates = @gate_class.where(feature_key: features.map(&:key)).pluck(:feature_key, :key, :value)
+          grouped_gates = gates.inject({}) do |hash, (feature_key, key, value)|
+            hash[feature_key] ||= []
+            hash[feature_key] << [key, value]
+            hash
+          end
+
+          result = {}
+          features.each do |feature|
+            result[feature.key] = result_for_gates(feature, grouped_gates[feature.key])
+          end
+          result
         end
-        result
       end
 
       def get_all
-        result = Hash.new { |hash, key| hash[key] = default_config }
+        with_connection(@feature_class) do
+          # query the gates from the db in a single query
+          features = ::Arel::Table.new(@feature_class.table_name.to_sym)
+          gates = ::Arel::Table.new(@gate_class.table_name.to_sym)
+          rows_query = features.join(gates, ::Arel::Nodes::OuterJoin)
+            .on(features[:key].eq(gates[:feature_key]))
+            .project(features[:key].as('feature_key'), gates[:key], gates[:value])
+          gates = @feature_class.connection.select_rows(rows_query)
 
-        @feature_class.includes(:gates).all.each do |f|
-          feature = Flipper::Feature.new(f.key, self)
-          result[feature.key] = result_for_feature(feature, f.gates)
+          # group the gates by feature key
+          grouped_gates = gates.inject({}) do |hash, (feature_key, key, value)|
+            hash[feature_key] ||= []
+            hash[feature_key] << [key, value]
+            hash
+          end
+
+          # build up the result hash
+          result = Hash.new { |hash, key| hash[key] = default_config }
+          features = grouped_gates.keys.map { |key| Flipper::Feature.new(key, self) }
+          features.each do |feature|
+            result[feature.key] = result_for_gates(feature, grouped_gates[feature.key])
+          end
+          result
         end
-
-        result
       end
 
       # Public: Enables a gate for a given thing.
@@ -165,7 +192,9 @@ module Flipper
         when :json
           delete(feature, gate)
         when :set
-          @gate_class.where(feature_key: feature.key, key: gate.key, value: thing.value).destroy_all
+          with_connection(@gate_class) do
+            @gate_class.where(feature_key: feature.key, key: gate.key, value: thing.value).destroy_all
+          end
         else
           unsupported_data_type gate.data_type
         end
@@ -186,18 +215,21 @@ module Flipper
 
         raise VALUE_TO_TEXT_WARNING if json_feature && value_not_text?
 
-        @gate_class.transaction do
-          clear(feature) if clear_feature
-          delete(feature, gate)
-          begin
-            @gate_class.create! do |g|
-              g.feature_key = feature.key
-              g.key = gate.key
-              g.value = json_feature ? JSON.dump(thing.value) : thing.value.to_s
+        with_connection(@gate_class) do
+          @gate_class.transaction do
+            clear(feature) if clear_feature
+            delete(feature, gate)
+            @gate_class.where(feature_key: feature.key, key: gate.key).destroy_all
+            begin
+              @gate_class.create! do |g|
+                g.feature_key = feature.key
+                g.key = gate.key
+                g.value = json_feature ? JSON.dump(thing.value) : thing.value.to_s
+              end
+            rescue ::ActiveRecord::RecordNotUnique
+              # assume this happened concurrently with the same thing and its fine
+              # see https://github.com/jnunemaker/flipper/issues/544
             end
-          rescue ::ActiveRecord::RecordNotUnique
-            # assume this happened concurrently with the same thing and its fine
-            # see https://github.com/jnunemaker/flipper/issues/544
           end
         end
 
@@ -209,10 +241,12 @@ module Flipper
       end
 
       def enable_multi(feature, gate, thing)
-        @gate_class.create! do |g|
-          g.feature_key = feature.key
-          g.key = gate.key
-          g.value = thing.value
+        with_connection(@gate_class) do
+          @gate_class.create! do |g|
+            g.feature_key = feature.key
+            g.key = gate.key
+            g.value = thing.value.to_s
+          end
         end
 
         nil
@@ -220,22 +254,22 @@ module Flipper
         # already added so no need move on with life
       end
 
-      def result_for_feature(feature, db_gates)
-        db_gates ||= []
+      def result_for_gates(feature, gates)
         result = {}
+        gates ||= []
         feature.gates.each do |gate|
           result[gate.key] =
             case gate.data_type
             when :boolean, :integer
-              if detected_db_gate = db_gates.detect { |db_gate| db_gate.key == gate.key.to_s }
-                detected_db_gate.value
+              if row = gates.detect { |key, value| !key.nil? && key.to_sym == gate.key }
+                row.last
               end
             when :json
-              if detected_db_gate = db_gates.detect { |db_gate| db_gate.key == gate.key.to_s }
-                JSON.parse(detected_db_gate.value)
+              if row = gates.detect { |key, value| !key.nil? && key.to_sym == gate.key }
+                JSON.parse(row.last)
               end
             when :set
-              db_gates.select { |db_gate| db_gate.key == gate.key.to_s }.map(&:value).to_set
+              gates.select { |key, value| !key.nil? && key.to_sym == gate.key }.map(&:last).to_set
             else
               unsupported_data_type gate.data_type
             end
@@ -247,6 +281,10 @@ module Flipper
       # See https://github.com/jnunemaker/flipper/pull/692
       def value_not_text?
         @gate_class.column_for_attribute(:value).type != :text
+      end
+
+      def with_connection(model = @feature_class, &block)
+        model.connection_pool.with_connection(&block)
       end
     end
   end