+++ lib/flipper/adapters/memory.rb
@@ -1,4 +1,4 @@
-require 'set'
+require 'concurrent/atomic/read_write_lock'
 
 module Flipper
   module Adapters
@@ -12,90 +12,97 @@ module Flipper
 
       # Public
       def initialize(source = nil)
-        @source = source || {}
+        @source = Hash.new.update(source || {})
         @name = :memory
+        @lock = Concurrent::ReadWriteLock.new
       end
 
       # Public: The set of known features.
       def features
-        @source.keys.to_set
+        @lock.with_read_lock { @source.keys }.to_set
       end
 
       # Public: Adds a feature to the set of known features.
       def add(feature)
-        @source[feature.key] ||= default_config
+        @lock.with_write_lock { @source[feature.key] ||= default_config }
         true
       end
 
       # Public: Removes a feature from the set of known features and clears
       # all the values for the feature.
       def remove(feature)
-        @source.delete(feature.key)
+        @lock.with_write_lock { @source.delete(feature.key) }
         true
       end
 
       # Public: Clears all the gate values for a feature.
       def clear(feature)
-        @source[feature.key] = default_config
+        @lock.with_write_lock { @source[feature.key] = default_config }
         true
       end
 
       # Public
       def get(feature)
-        @source[feature.key] || default_config
+        @lock.with_read_lock { @source[feature.key] } || default_config
       end
 
       def get_multi(features)
-        result = {}
-        features.each do |feature|
-          result[feature.key] = @source[feature.key] || default_config
+        @lock.with_read_lock do
+          result = {}
+          features.each do |feature|
+            result[feature.key] = @source[feature.key] || default_config
+          end
+          result
         end
-        result
       end
 
       def get_all
-        @source
+        @lock.with_read_lock { @source.to_h }
       end
 
       # Public
       def enable(feature, gate, thing)
-        @source[feature.key] ||= default_config
+        @lock.with_write_lock do
+          @source[feature.key] ||= default_config
 
-        case gate.data_type
-        when :boolean
-          clear(feature)
-          @source[feature.key][gate.key] = thing.value.to_s
-        when :integer
-          @source[feature.key][gate.key] = thing.value.to_s
-        when :set
-          @source[feature.key][gate.key] << thing.value.to_s
-        when :json
-          @source[feature.key][gate.key] = thing.value
-        else
-          raise "#{gate} is not supported by this adapter yet"
+          case gate.data_type
+          when :boolean
+            @source[feature.key] = default_config
+            @source[feature.key][gate.key] = thing.value.to_s
+          when :integer
+            @source[feature.key][gate.key] = thing.value.to_s
+          when :set
+            @source[feature.key][gate.key] << thing.value.to_s
+          when :json
+            @source[feature.key][gate.key] = thing.value
+          else
+            raise "#{gate} is not supported by this adapter yet"
+          end
+
+          true
         end
-
-        true
       end
 
       # Public
       def disable(feature, gate, thing)
-        @source[feature.key] ||= default_config
+        @lock.with_write_lock do
+          @source[feature.key] ||= default_config
 
-        case gate.data_type
-        when :boolean
-          clear(feature)
-        when :integer
-          @source[feature.key][gate.key] = thing.value.to_s
-        when :set
-          @source[feature.key][gate.key].delete thing.value.to_s
-        when :json
-          @source[feature.key].delete(gate.key)
-        else
-          raise "#{gate} is not supported by this adapter yet"
+          case gate.data_type
+          when :boolean
+            @source[feature.key] = default_config
+          when :integer
+            @source[feature.key][gate.key] = thing.value.to_s
+          when :set
+            @source[feature.key][gate.key].delete thing.value.to_s
+          when :json
+            @source[feature.key].delete(gate.key)
+          else
+            raise "#{gate} is not supported by this adapter yet"
+          end
+
+          true
         end
-
-        true
       end
 
       # Public
@@ -105,6 +112,12 @@ module Flipper
           "source=#{@source.inspect}",
         ]
         "#<#{self.class.name}:#{object_id} #{attributes.join(', ')}>"
+      end
+
+      # Public: a more efficient implementation of import for this adapter
+      def import(source_adapter)
+        get_all = source_adapter.get_all
+        @lock.with_write_lock { @source.replace(get_all) }
       end
     end
   end