+++ lib/flipper/gate_values.rb
@@ -3,17 +3,6 @@ require 'flipper/typecast'
 
 module Flipper
   class GateValues
-    # Private: Array of instance variables that are readable through the []
-    # instance method.
-    LegitIvars = {
-      'boolean' => '@boolean',
-      'actors' => '@actors',
-      'groups' => '@groups',
-      'expression' => '@expression',
-      'percentage_of_time' => '@percentage_of_time',
-      'percentage_of_actors' => '@percentage_of_actors',
-    }.freeze
-
     attr_reader :boolean
     attr_reader :actors
     attr_reader :groups
@@ -28,12 +17,6 @@ module Flipper
       @expression = adapter_values[:expression]
       @percentage_of_actors = Typecast.to_percentage(adapter_values[:percentage_of_actors])
       @percentage_of_time = Typecast.to_percentage(adapter_values[:percentage_of_time])
-    end
-
-    def [](key)
-      if ivar = LegitIvars[key.to_s]
-        instance_variable_get(ivar)
-      end
     end
 
     def eql?(other)