+++ spec/flipper/adapters/rollout_spec.rb
@@ -11,10 +11,8 @@ RSpec.describe Flipper::Adapters::Rollout do
   let(:destination_flipper) { Flipper.new(destination_adapter) }
 
   before do
-    begin
+    skip_on_error(Redis::CannotConnectError, 'Redis not available') do
       redis.flushdb
-    rescue Redis::CannotConnectError
-      ENV['CI'] ? raise : skip('Redis not available')
     end
   end
 
@@ -120,7 +118,7 @@ RSpec.describe Flipper::Adapters::Rollout do
     expect(feature.boolean_value).to eq(false)
     expect(feature.actors_value).to eq(Set.new)
     expect(feature.groups_value).to eq(Set.new)
-    expect(feature.percentage_of_actors_value).to be(25.0)
+    expect(feature.percentage_of_actors_value).to be(25)
 
     feature = destination_flipper[:verbose_logging]
     expect(feature.boolean_value).to eq(false)